refactor(ui): remove barrel files and colocate prop types
Delete index.ts re-export barrels for ShowMoreReplies, ThreadExpandToggle,
ThreadTree, ReplyWithNested, useNestedReplies, useReplyStream, and
useThreadReplies. Inline prop interfaces into their component files
and update all import paths to reference source modules directly.

Also simplifies the ShowMoreReplies connector layout by unifying the
duplicated isLast/regular JSX branches and corrects the persistNewStreamChunk
JSDoc to match its actual behavior.

Refs: #1542
