--- src/components/molecules/ShowMoreReplies/index.ts
@@ -1,2 +1,0 @@
-export * from './ShowMoreReplies';
-export * from './ShowMoreReplies.types';
