--- src/components/molecules/ThreadExpandToggle/index.ts
@@ -1,1 +1,0 @@
-export * from './ThreadExpandToggle';
