--- src/components/organisms/ReplyWithNested/index.ts
@@ -1,2 +1,0 @@
-export * from './ReplyWithNested';
-export * from './ReplyWithNested.types';
