--- src/components/organisms/ThreadTree/ThreadTree.types.ts
@@ -1,6 +1,0 @@
-export interface ThreadTreeProps {
-  /** The composite post ID of the parent (Level 0) post */
-  postId: string;
-  /** Whether to show the QuickReply input at the end */
-  showQuickReply?: boolean;
-}
