--- src/components/organisms/ThreadTree/index.ts
@@ -1,2 +1,0 @@
-export * from './ThreadTree';
-export * from './ThreadTree.types';
