+++ src/hooks/index.ts
@@ -59,7 +59,7 @@ export * from './usePostTags';
 export * from './usePostTaggers';
 export * from './usePostParticipants';
 export * from './usePostCounts';
-export * from './useNestedReplies';
+export * from './useNestedReplies/useNestedReplies';
 export * from './useEntityTags';
 export * from './useStickyWhenFits';
 export * from './useSearchInput';
@@ -85,8 +85,8 @@ export * from './useRelativeTime';
 export * from './useInviteCodeSignUp';
 export * from './useEnrichedTags';
 export * from './usePullToRefresh';
-export * from './useReplyStream';
-export * from './useThreadReplies';
+export * from './useReplyStream/useReplyStream';
+export * from './useThreadReplies/useThreadReplies';
 export * from './useKeyboardVisible';
 export * from './useKeyboardOffset';
 export * from './useSignOut';
