--- src/hooks/useNestedReplies/index.ts
@@ -1,3 +1,0 @@
-export * from './useNestedReplies';
-export * from './useNestedReplies.types';
-export * from './useNestedReplies.constants';
