+++ src/hooks/useNestedReplies/useNestedReplies.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 // Direct import to avoid circular dependency (this hook is exported from @/hooks)
-import { useReplyStream } from '@/hooks/useReplyStream';
+import { useReplyStream } from '@/hooks/useReplyStream/useReplyStream';
 import { DEFAULT_MAX_NESTED, DEFAULT_MAX_DEPTH } from './useNestedReplies.constants';
 import type { UseNestedRepliesOptions, UseNestedRepliesResult } from './useNestedReplies.types';
