--- src/hooks/useThreadReplies/index.ts
@@ -1,3 +1,0 @@
-export * from './useThreadReplies';
-export * from './useThreadReplies.types';
-export * from './useThreadReplies.constants';
